requirements.txt
python-dotenv==1.1.0
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.9.1
google-generativeai==0.8.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
